# src/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Social Network for Music API",
    "description": "API del sito web Social Network for Music (PWM_SNM)",
    "version": "1.0.0"
  },
  "host": "localhost:3100",
  "basePath": "/",
  "tags": [
    {
      "name": "artists",
      "description": "Route legate agli artisti"
    },
    {
      "name": "auth",
      "description": "Route legate all'autenticazione degli utenti"
    },
    {
      "name": "genres",
      "description": "Route legate ai generi musicali"
    },
    {
      "name": "playlists",
      "description": "Route legate alle playlist (pubbliche o private che siano)"
    },
    {
      "name": "songs",
      "description": "Route legate alle canzoni"
    },
    {
      "name": "user",
      "description": "Route legate agli utenti dell'applicativo"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Registra un utente alla piattaforma",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "User registration Data",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": "username"
                },
                "name": {
                  "type": "string",
                  "example": "name"
                },
                "surname": {
                  "type": "string",
                  "example": "surname"
                },
                "email": {
                  "type": "string",
                  "example": "email@it"
                },
                "password": {
                  "type": "string",
                  "example": "pwd"
                },
                "password1": {
                  "type": "string",
                  "example": "pwd"
                },
                "favouriteArtists": {
                  "type": "array",
                  "example": [],
                  "items": {}
                },
                "favouriteGenres": {
                  "type": "array",
                  "example": [],
                  "items": {}
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Manca uno qualsiasi dei parametri che dovrebbero essere nel body. Inoltre password e password1 dovrebbero essere uguali."
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Fa eseguire l'accesso a SNM ad un utente",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "User login Data",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": "username/email"
                },
                "password": {
                  "type": "string",
                  "example": "pwd"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Mancano o l'username o la password (Oppure è stata inserita una combinazione errata)"
          }
        }
      }
    },
    "/api/users/{id}": {
      "delete": {
        "tags": [
          "user"
        ],
        "summary": "Elimino un utente specifico",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Username o email dell'utente da eliminare."
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Viene restituito un singolo utente",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Username o email dell'utente di cui si vogliono i dati."
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "user"
        ],
        "summary": "Viene modificato il profilo di un utente",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Username o email dell'utente da modificare."
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Request body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "currentPassword": {
                  "type": "string",
                  "example": "current"
                },
                "newPassword": {
                  "type": "string",
                  "example": "newPsw"
                },
                "newPassword1": {
                  "type": "string",
                  "example": "newPsw"
                },
                "favouriteGenres": {
                  "type": "array",
                  "example": [],
                  "items": {}
                },
                "favouriteArtists": {
                  "type": "array",
                  "example": [],
                  "items": {}
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/users/": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Viene restituito un array di utenti",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/artists/": {
      "get": {
        "tags": [
          "artists"
        ],
        "summary": "Viene restituito un array che contiene tutti gli artisti memorizzati nel database",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/genres/": {
      "get": {
        "tags": [
          "genres"
        ],
        "summary": "Viene restituito un array che contiene tutti i generi musicali memorizzati nel database",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/songs/": {
      "get": {
        "tags": [
          "songs"
        ],
        "summary": "Viene restituito un array di canzoni",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/songs/{id}": {
      "get": {
        "tags": [
          "songs"
        ],
        "summary": "Viene restituita un'unica canzone con un certo id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Song ID."
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/playlist/{owner}/{playlistName}": {
      "get": {
        "tags": [
          "playlists"
        ],
        "summary": "Viene restituita una singola playlist (identificata da owner e nome della playlist)",
        "description": "",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Username del proprietario della playlist"
          },
          {
            "name": "playlistName",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Nome della playlist"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/playlist/{owner}": {
      "get": {
        "tags": [
          "playlists"
        ],
        "summary": "Vengono restituite tutte le playlist dell",
        "description": "",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Username del proprietario della playlist"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/playlist/{owner}/{name}": {
      "put": {
        "tags": [
          "playlists"
        ],
        "summary": "Modifico i dati di una singola playlist",
        "description": "",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Username del proprietario della playlist"
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Vecchio nome della playlist"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Request body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "PlaylistName"
                },
                "songs": {
                  "type": "array",
                  "example": [],
                  "items": {}
                },
                "tags": {
                  "type": "string",
                  "example": "#Dei#Tag#A#Caso"
                },
                "privacy": {
                  "type": "string",
                  "example": "public"
                },
                "description": {
                  "type": "string",
                  "example": "Descrizione della playlist"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "playlists"
        ],
        "summary": "Elimino una playlist specifica",
        "description": "",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Username del proprietario della playlist"
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Nome della playlist"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/playlist/": {
      "post": {
        "tags": [
          "playlists"
        ],
        "summary": "Inserisco una nuova playlist",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Request body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "PlaylistName"
                },
                "owner": {
                  "type": "string",
                  "example": "ownerUsername"
                },
                "songs": {
                  "type": "array",
                  "example": [],
                  "items": {}
                },
                "tags": {
                  "type": "string",
                  "example": "#Dei#Tag#A#Caso"
                },
                "privacy": {
                  "type": "string",
                  "example": "public"
                },
                "description": {
                  "type": "string",
                  "example": "Descrizione della playlist"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/publicPlaylists/": {
      "get": {
        "tags": [
          "playlists"
        ],
        "summary": "Vengono restituite tutte le playlist pubbliche che sono all",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}